basic things. What needs to be done now is to know how to configure the minimum settings that we have, and then how to call the correct libraries in the code. The goal is to get it read in via SPI and wit th the timer, send it to a software buffer, and then have a second timer to pad the serial messages and send it out via UART/USB asap.
What I currently don';t understand is why the system is not committing cometpltel;y. There is a website that explains it well. The issue is to do with the configs itself. So pay more attention to that tomorrow. Using a new plotter from python shouldn't take too long either